+++ x/oracle/client/cli/query.go
@@ -22,6 +22,7 @@ func GetQueryCmd() *cobra.Command {
 
 	cmd.AddCommand(
 		QueryParamsCmd(),
+		QueryInturnRelayerCmd(),
 	)
 
 	return cmd
@@ -50,6 +51,37 @@ $ <appd> query oracle params
 			}
 
 			return clientCtx.PrintProto(&res.Params)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// QueryParamsCmd returns the command handler for evidence parameter querying.
+func QueryInturnRelayerCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "inturn-relayer",
+		Short: "Query the inturn relayer",
+		Args:  cobra.NoArgs,
+		Long: strings.TrimSpace(`Query the inturn relayer:
+
+$ <appd> query oracle inturn-relayer
+`),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			queryClient := types.NewQueryClient(clientCtx)
+			res, err := queryClient.InturnRelayer(cmd.Context(), &types.QueryInturnRelayerRequest{})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
 		},
 	}