+++ x/params/proposal_handler.go
@@ -24,8 +24,6 @@ func NewParamChangeProposalHandler(k keeper.Keeper) govtypes.Handler {
 }
 
 func handleParameterChangeProposal(ctx sdk.Context, k keeper.Keeper, p *proposal.ParameterChangeProposal) error {
-	// for govv1.MsgExecLegacyContent, validation is applied by writing to cache Subspace store, since gnfd does not hold
-	// the subspace for crosschain parameters. the validation is skipped.
 	if p.CrossChain {
 		return k.SyncParams(ctx, p)
 	}