+++ x/staking/types/msg_test.go
@@ -43,7 +43,8 @@ func TestMsgDecode(t *testing.T) {
 	msg, err := types.NewMsgCreateValidator(
 		valAddr1, pk1,
 		coinPos, types.Description{}, commission1, sdk.OneInt(),
-		sdk.AccAddress(valAddr1), sdk.AccAddress(valAddr1), sdk.AccAddress(valAddr1), blsPk,
+		sdk.AccAddress(valAddr1), sdk.AccAddress(valAddr1),
+		sdk.AccAddress(valAddr1), sdk.AccAddress(valAddr1), blsPk,
 	)
 	require.NoError(t, err)
 	msgSerialized, err := cdc.MarshalInterface(msg)
@@ -91,7 +92,8 @@ func TestMsgCreateValidator(t *testing.T) {
 		msg, err := types.NewMsgCreateValidator(
 			tc.validatorAddr, tc.pubkey,
 			tc.bond, description, tc.CommissionRates, tc.minSelfDelegation,
-			sdk.AccAddress(tc.validatorAddr), sdk.AccAddress(tc.validatorAddr), sdk.AccAddress(tc.validatorAddr), blsPk,
+			sdk.AccAddress(tc.validatorAddr), sdk.AccAddress(tc.validatorAddr),
+			sdk.AccAddress(tc.validatorAddr), sdk.AccAddress(tc.validatorAddr), blsPk,
 		)
 		require.NoError(t, err)
 		if tc.expectPass {
@@ -125,7 +127,7 @@ func TestMsgEditValidator(t *testing.T) {
 
 		msg := types.NewMsgEditValidator(
 			tc.validatorAddr, description, &newRate, &tc.minSelfDelegation,
-			sdk.AccAddress(tc.validatorAddr), blsPk)
+			sdk.AccAddress(tc.validatorAddr), sdk.AccAddress(tc.validatorAddr), blsPk)
 		if tc.expectPass {
 			require.Nil(t, msg.ValidateBasic(), "test: %v", tc.name)
 		} else {